requirements/base.txt
Django>=5.0,<5.1
djangorestframework>=3.14,<3.15
djangorestframework-simplejwt>=5.3,<5.4
drf-spectacular>=0.27,<0.28
django-filter>=23.5,<24.0
django-cors-headers>=4.3,<4.4
django-environ>=0.11,<0.12
django-storages>=1.14,<1.15
boto3>=1.26,<2.0
Pillow>=10.0,<11.0
psycopg2-binary>=2.9,<3.0
